requirements.txt
pandas
requests
ipykernel
pydantic
loguru
python-dotenv
tqdm
torch
torchvision
pycocotools
